+++ sentry-android-okhttp/src/test/java/io/sentry/android/okhttp/SentryOkHttpEventTest.kt
@@ -1,6 +1,7 @@
 package io.sentry.android.okhttp
 
 import io.sentry.Breadcrumb
+import io.sentry.Hint
 import io.sentry.IHub
 import io.sentry.ISentryExecutorService
 import io.sentry.ISpan
@@ -21,6 +22,7 @@ import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.REQUEST_HEAD
 import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_BODY_EVENT
 import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_HEADERS_EVENT
 import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.SECURE_CONNECT_EVENT
+import io.sentry.exception.SentryHttpClientException
 import io.sentry.test.getProperty
 import okhttp3.Protocol
 import okhttp3.Request
@@ -28,6 +30,7 @@ import okhttp3.Response
 import okhttp3.mockwebserver.MockWebServer
 import org.mockito.kotlin.any
 import org.mockito.kotlin.anyOrNull
+import org.mockito.kotlin.argThat
 import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.check
 import org.mockito.kotlin.eq
@@ -38,6 +41,7 @@ import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import java.util.concurrent.Future
 import java.util.concurrent.RejectedExecutionException
+import kotlin.RuntimeException
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -482,6 +486,7 @@ class SentryOkHttpEventTest {
         sut.finishSpan(REQUEST_HEADERS_EVENT) { it.status = SpanStatus.INTERNAL_ERROR }
         sut.finishSpan("random event") { it.status = SpanStatus.DEADLINE_EXCEEDED }
         sut.finishSpan(CONNECTION_EVENT)
+        sut.finishEvent()
         val spans = sut.getEventSpans()
         val connectionSpan = spans[CONNECTION_EVENT] as Span?
         val requestHeadersSpan = spans[REQUEST_HEADERS_EVENT] as Span?
@@ -497,6 +502,34 @@ class SentryOkHttpEventTest {
         assertEquals(SpanStatus.INTERNAL_ERROR, connectionSpan.status)
         // random event was finished last with DEADLINE_EXCEEDED, and it propagates to root call
         assertEquals(SpanStatus.DEADLINE_EXCEEDED, sut.callRootSpan!!.status)
+    }
+
+    @Test
+    fun `finishEvent moves throwables from inner span to call root span`() {
+        val sut = fixture.getSut()
+        val throwable = RuntimeException()
+        sut.startSpan(CONNECTION_EVENT)
+        sut.startSpan("random event")
+        sut.finishSpan("random event") { it.status = SpanStatus.DEADLINE_EXCEEDED }
+        sut.finishSpan(CONNECTION_EVENT) {
+            it.status = SpanStatus.INTERNAL_ERROR
+            it.throwable = throwable
+        }
+        sut.finishEvent()
+        val spans = sut.getEventSpans()
+        val connectionSpan = spans[CONNECTION_EVENT] as Span?
+        val randomEventSpan = spans["random event"] as Span?
+        assertNotNull(connectionSpan)
+        assertNotNull(randomEventSpan)
+        // randomEventSpan was finished with DEADLINE_EXCEEDED
+        assertEquals(SpanStatus.DEADLINE_EXCEEDED, randomEventSpan.status)
+        // connectionSpan was finished with INTERNAL_ERROR
+        assertEquals(SpanStatus.INTERNAL_ERROR, connectionSpan.status)
+
+        // connectionSpan was finished last with INTERNAL_ERROR and a throwable, and it's moved to the root call
+        assertEquals(SpanStatus.INTERNAL_ERROR, sut.callRootSpan!!.status)
+        assertEquals(throwable, sut.callRootSpan.throwable)
+        assertNull(connectionSpan.throwable)
     }
 
     @Test
@@ -529,6 +562,33 @@ class SentryOkHttpEventTest {
         whenever(fixture.hub.options).thenReturn(SentryOptions().apply { this.executorService = executorService })
         val sut = fixture.getSut()
         sut.scheduleFinish(mock())
+    }
+
+    @Test
+    fun `setClientErrorResponse will capture the client error on finishEvent`() {
+        val sut = fixture.getSut()
+        val clientErrorResponse = mock<Response>()
+        whenever(clientErrorResponse.request).thenReturn(fixture.mockRequest)
+        sut.setClientErrorResponse(clientErrorResponse)
+        verify(fixture.hub, never()).captureEvent(any(), any<Hint>())
+        sut.finishEvent()
+        verify(fixture.hub).captureEvent(
+            argThat {
+                throwable is SentryHttpClientException &&
+                    throwable!!.message!!.startsWith("HTTP Client Error with status code: ")
+            },
+            argThat<Hint> {
+                get(TypeCheckHint.OKHTTP_REQUEST) != null &&
+                    get(TypeCheckHint.OKHTTP_RESPONSE) != null
+            }
+        )
+    }
+
+    @Test
+    fun `when setClientErrorResponse is not called, no client error is captured`() {
+        val sut = fixture.getSut()
+        sut.finishEvent()
+        verify(fixture.hub, never()).captureEvent(any(), any<Hint>())
     }
 
     /** Retrieve all the spans started in the event using reflection. */